cli: a listing command refuses a name it cannot answer about
Measured on the built binary: "tfg formats png svg" described png, said
nothing at all about svg and ended with zero, and "tfg preset list some-name"
printed the whole list and ignored the name though that operation takes no
name at all. Both are silence about an argument somebody typed, and in a
script a confident answer about the wrong thing is worse than a refusal.

The second one is not in the observation that asked for this. It came out of
measuring the first, and fixing only the named half would have left that
observation's own sentence true about a different pair of commands.

Both now end with 2 and name the word they could not use. The guard asks four
commands at once, including the two that already refused, and requires the
refusal to name the rejected word - an exit code alone is satisfied by a build
that refuses everything, so the other half checks these commands still answer.

The flag spelling guard stops naming its directories by hand and derives them
from the layer map. internal/damage arrived on layer 2, under both surfaces,
without joining that list, and the hole would have taken every later package
too. The layer map is the right source because the layering guard already
refuses a package that is neither on the ladder nor declared test only, and
because it leaves out internal/oracle on purpose - that one carries real
command lines for inkscape and ffprobe in raw strings. 113 files scanned
became 128, with zero findings.

Refusing the extra name pushed internal/cli/preset.go past the file ceiling,
so it is split by what the parts do: presetcmd.go answers somebody who typed
tfg preset, preset.go turns a named question into targets. The ratchet then
took the ceiling down, 433 to 408 and crowdedFiles 1 to 0.

Three new mutations, and four existing ones retargeted after the split - only
staleness.py said they had gone stale. All eight run and caught.
